refactor: deepen app runtime composition
